Refactored getVehicleStats function into its own file to debug integration tests
