+++ client/views/select-vehicle.html
@@ -45,7 +45,7 @@
             <p><strong>Range:</strong> <span id="car-range"> --- </span> miles</p>
             <p><strong>Manufacturer:</strong> <span id="car-manufacturer"> --- </span></p>
  
-            <button type="button" onclick="window.location.href='map.html'" class="btn btn-dark w-100">Submit</button>
+            <button id='vehicle-selection-button' type="button" onclick="window.location.href='map.html'" class="btn btn-dark w-100">Submit</button>
         </div>
     </main>
